+++ CHANGELOG.md
@@ -6,6 +6,11 @@
   - Example: 10.2.1.4 is the 5th version that supports khiops 10.2.1.
 - Internals: Changes in *Internals* sections are unlikely to be of interest for data scientists.
 
+## Unreleased
+
+### Fixed
+- (`core`) Fix a RuntimeException error when automatically correcting deprecated data paths in task args
+
 ## 11.0.1.0 - 2026-07-02
 
 ### Added